backend/requirements.txt
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
babel==2.18.0
blinker==1.9.0
certifi==2026.4.22
chardet==7.4.3
charset-normalizer==3.4.7
click==8.3.3
click-plugins==1.1.1.2
cligj==0.7.2
contourpy==1.3.3
cycler==0.12.1
distinctipy==1.3.4
docutils==0.22.4
fastapi==0.136.1
fiona==1.10.1
Flask==3.1.3
flask-cors==6.0.2
flexcache==0.3
flexparser==0.4
fonttools==4.62.1
frictionless==5.19.0
frozenlist==1.8.0
fsspec==2026.3.0
GDAL==3.12.4
geometamaker==0.3.1
geopandas==1.1.3
h11==0.16.0
humanize==4.15.0
idna==3.13
importlib_metadata==9.0.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
matplotlib==3.10.9
mdurl==0.1.2
multidict==6.7.1
narwhals==2.20.0
natcap.invest==3.19.0
numpy==2.4.4
packaging==26.2
pandas==3.0.2
petl==1.7.17
pillow==12.2.0
Pint==0.25.3
platformdirs==4.9.6
propcache==0.4.1
psutil==7.2.2
pydantic==2.13.3
pydantic_core==2.46.3
pygeoprocessing==2.4.11
Pygments==2.20.0
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.2
Pyro5==5.16
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
PyYAML==6.0.3
rasterio==1.5.0
referencing==0.37.0
requests==2.33.1
retrying==1.4.2
rfc3986==2.0.0
rich==15.0.0
rpds-py==0.30.0
rtree==1.4.1
scipy==1.17.1
serpent==1.42
shapely==2.1.2
shellingham==1.5.4
simpleeval==1.0.7
six==1.17.0
starlette==1.0.0
tabulate==0.10.0
taskgraph==0.11.2
text-unidecode==1.3
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
validators==0.35.0
Werkzeug==3.1.8
yarl==1.23.0
zipp==3.23.1
